@@ -69,7 +69,7 @@ class User < ApplicationRecord
     validate_length(:password, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR6"), 6, 128)
     validate_password(:password, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR5")) unless self.password.blank?
     validate_mail_address(:email, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR7")) unless self.email.blank?
-    validate_date(:birth_date, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR8")) unless self.birth_date.blank?
+    validate_date(:birth_date, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR8")) unless self.birth_date_before_type_cast.blank?
     validate_presence(:sex_cd, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR10")) if self.sex_cd.blank?
     #validate_name_no_prefix(:name_no_prefix, I18n.t("admin.user.PRI_ADM_USR_REGISTERUST_ERROR11")) unless self.name_no_prefix.blank?
 
@@ -150,7 +150,7 @@ class User < ApplicationRecord
     if self.errors.messages.count > 0
       result = false
     else
-      update_attributes(params, *options)
+      update(params, *options)
       result = true
     end
 
@@ -224,7 +224,7 @@ class User < ApplicationRecord
     if self.errors.messages.count > 0
       result = false
     else
-      result = update_attributes(email: params[:new_email])
+      result = update(email: params[:new_email])
     end
 
     result
