@@ -22,6 +22,11 @@
 #++
 
 class ApplicationController < ActionController::Base
+
+  class Forbidden < ActionController::ActionControllerError; end
+
+  VIEW_COUSE_ORER = "school_year DESC, season_cd, day_cd, hour_cd"
+
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
@@ -33,12 +38,14 @@ class ApplicationController < ActionController::Base
   before_action :set_locale
   before_action :check_permission
 
+  rescue_from Exception, with: :render_500
+  rescue_from Forbidden, with: :render_403
   rescue_from ActiveRecord::RecordNotFound, with: :render_404
   rescue_from ActionController::RoutingError, with: :render_404
-  rescue_from Exception, with: :render_500
 
-  NOT_ASSIGNED = "not assigned"
-  NOT_ENROLLED = "not enrolled"
+  def render_403
+    render template: 'errors/error500', status: :forbidden
+  end
 
   def render_404
     render template: 'errors/error404', status: :not_found
@@ -56,6 +63,9 @@ class ApplicationController < ActionController::Base
   end
 
   def other_courses
+    @list_controller_name = params[:controller_name]
+    @list_action = params[:list_action]
+
     set_course
     set_other_courses
     render :partial => "shared/courses/course_list"
@@ -84,23 +94,21 @@ class ApplicationController < ActionController::Base
     def set_courses
       if current_user.admin?
         if params[:course_name].blank?
-          @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").page(params[:page])
+          @courses = Course.order(VIEW_COUSE_ORER).page(params[:page])
         else
-          @courses = Course.where("course_name LIKE ?", "%#{params[:course_name]}%").order("school_year DESC, day_cd, hour_cd, season_cd").page(params[:page])
+          @courses = Course.where("course_name LIKE ?", "%#{params[:course_name]}%").order(VIEW_COUSE_ORER).page(params[:page])
         end
-#          @courses = Course.all.order("school_year DESC, day_cd, hour_cd, season_cd").page(params[:page])
       elsif current_user.teacher?
         if params[:course_name].blank?
-          @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_assigned_users).where("user_id = ? AND indirect_use_flag = ? AND (courses.term_flag = ? OR courseware_flag = ?)", current_user.id, false, true, true).page(params[:page])
+          @courses = Course.order(VIEW_COUSE_ORER).joins(:course_assigned_users).where("user_id = ? AND indirect_use_flag = ? AND (courses.term_flag = ? OR courseware_flag = ?)", current_user.id, false, true, true).page(params[:page])
         else
-          @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_assigned_users).where("user_id = ? AND indirect_use_flag = ? AND (courses.term_flag = ? OR courseware_flag = ?) AND course_name LIKE ?", current_user.id, false, true, true, "%#{params[:course_name]}%").page(params[:page])
+          @courses = Course.order(VIEW_COUSE_ORER).joins(:course_assigned_users).where("user_id = ? AND indirect_use_flag = ? AND (courses.term_flag = ? OR courseware_flag = ?) AND course_name LIKE ?", current_user.id, false, true, true, "%#{params[:course_name]}%").page(params[:page])
         end
       else
-#        @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_enrollment_users).where("user_id = ? AND indirect_use_flag = ? AND courses.term_flag = ? AND courseware_flag = ? AND course_name LIKE ?", current_user.id, false, true, true, "%#{params[:course_name]}%").page(params[:page])
         if params[:course_name].blank?
-          @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_enrollment_users).where("user_id = ? AND indirect_use_flag = ? AND courses.term_flag = ? AND courseware_flag = ?", current_user.id, false, true, false).page(params[:page])
+          @courses = Course.order(VIEW_COUSE_ORER).joins(:course_enrollment_users).where("user_id = ? AND indirect_use_flag = ? AND courses.term_flag = ? AND courseware_flag = ?", current_user.id, false, true, false).page(params[:page])
         else
-          @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_enrollment_users).where("user_id = ? AND indirect_use_flag = ? AND courses.term_flag = ? AND courseware_flag = ? AND course_name LIKE ?", current_user.id, false, true, false, "%#{params[:course_name]}%").page(params[:page])
+          @courses = Course.order(VIEW_COUSE_ORER).joins(:course_enrollment_users).where("user_id = ? AND indirect_use_flag = ? AND courses.term_flag = ? AND courseware_flag = ? AND course_name LIKE ?", current_user.id, false, true, false, "%#{params[:course_name]}%").page(params[:page])
         end
       end
     end
@@ -109,13 +117,13 @@ class ApplicationController < ActionController::Base
       if current_user.admin?
         if params[:course_name].blank?
           @courses = Course.where("courses.id != ?", params[:course_id]).
-            order("school_year DESC, day_cd, hour_cd, season_cd").page(params[:page])
+            order(VIEW_COUSE_ORER).page(params[:page])
         else
           @courses = Course.where("courses.id != ? AND course_name LIKE ?", params[:course_id], "%#{params[:course_name]}%").
-            order("school_year DESC, day_cd, hour_cd, season_cd").page(params[:page])
+            order(VIEW_COUSE_ORER).page(params[:page])
         end
       elsif current_user.teacher?
-        @courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").joins(:course_assigned_users).
+        @courses = Course.order(VIEW_COUSE_ORER).joins(:course_assigned_users).
           where("user_id = ? AND indirect_use_flag = ? AND (courses.term_flag = ? OR courseware_flag = ?) AND courses.id != ? AND course_name LIKE ?",
             current_user.id, false, true, true, params[:course_id], "%#{params[:course_name]}%")
           .page(params[:page])
@@ -190,7 +198,9 @@ class ApplicationController < ActionController::Base
         assigned = course.course_assigned_users.where(user_id: current_user.id).first
       end
       
-      raise NOT_ASSIGNED if assigned.blank?
+      if assigned.blank?
+        raise Forbidden
+      end
     end
 
     def require_enrolled
@@ -198,13 +208,15 @@ class ApplicationController < ActionController::Base
 
       course = get_require_course
       if current_user.teacher?
-        assigned = course.course_assigned_users.where(user_id: current_user.id).first
-        raise NOT_ASSIGNED if assigned.blank?
+        enrolled = course.course_assigned_users.where(user_id: current_user.id).first
 
       else
         enrolled = course.course_enrollment_users.where(user_id: current_user.id).first
-        raise NOT_ENROLLED if enrolled.blank?
 
+      end
+
+      if enrolled.blank?
+        raise Forbidden
       end
     end
 
@@ -221,7 +233,10 @@ class ApplicationController < ActionController::Base
       if assigned.blank? && course.open_course_flag == Settings.COURSE_OPENCOURSEFLG_PUBLIC
         assigned = course.open_course_assigned_users.where(user_id: current_user.id).first
       end
-      raise NOT_ENROLLED if assigned.blank?
+
+      if assigned.blank?
+        raise Forbidden
+      end
     end
 
     def get_content_type(file_name)
