@@ -43,6 +43,7 @@ class Teacher::CompoundsController < ApplicationController
 
   def create
     @generic_page = Compound.new(generic_page_params)
+    @generic_page.course_id = params[:course_id]
     unless params[:generic_page][:file].blank?
       @generic_page.file = params[:generic_page][:file]
     end
@@ -89,11 +90,11 @@ class Teacher::CompoundsController < ApplicationController
 
   def select_course
     if current_user.admin?
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         where("courses.id != ? AND course_name LIKE ?", params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
     else
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         joins(:course_assigned_users).
         where("user_id = ? AND courses.id != ? AND course_name LIKE ?", current_user.id, params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
@@ -165,8 +166,8 @@ class Teacher::CompoundsController < ApplicationController
 
     def generic_page_params
       params.require(:generic_page).permit(:course_id, :type_cd, :generic_page_title, :max_count,
-        :pass_grade, :upload_flag, :file, :start_pass, :start_time,
-        :end_time, :self_type, :self_pass, :material_memo)
+        :pass_grade, :file, :start_pass, :start_time, :end_time, :self_type,
+        :correct_answer_display_flag, :self_pass, :material_memo)
     end
 
     def material_file_params
