@@ -76,11 +76,11 @@ class Teacher::EssaysController < ApplicationController
 
   def select_course
     if current_user.admin?
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         where("courses.id != ? AND course_name LIKE ?", params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
     else
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         joins(:course_assigned_users).
         where("user_id = ? AND courses.id != ? AND course_name LIKE ?", current_user.id, params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
@@ -160,7 +160,7 @@ class Teacher::EssaysController < ApplicationController
     end
 
     def generic_page_params
-      params.require(:generic_page).permit(:course_id, :type_cd, :generic_page_title, :upload_flag,
+      params.require(:generic_page).permit(:course_id, :type_cd, :generic_page_title,
         :file, :answer_file, :assignment_essay_return_method_cd, :score_open_flag,
         :start_pass, :start_time, :end_time, :pre_grading_enable_flag, :material_memo)
     end
