@@ -35,7 +35,6 @@ class Teacher::QuestionnairesController < ApplicationController
     @generic_page = GenericPage.new()
     @generic_page.course = @course
     @generic_page.type_cd = Settings.GENERICPAGE_TYPECD_QUESTIONNAIRECODE
-    @generic_page.upload_flag = GenericPage::TYPE_NOFILEUPLOAD
     @generic_page.edit_flag = 0
     @generic_page.anonymous_flag = 0
   end
@@ -72,11 +71,11 @@ class Teacher::QuestionnairesController < ApplicationController
 
   def select_course
     if current_user.admin?
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         where("courses.id != ? AND course_name LIKE ?", params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
     else
-      @other_courses = Course.order("school_year DESC, day_cd, hour_cd, season_cd").
+      @other_courses = Course.order(VIEW_COUSE_ORER).
         joins(:course_assigned_users).
         where("user_id = ? AND courses.id != ? AND course_name LIKE ?", current_user.id, params[:course_id], "%#{params[:course_name]}%").
         page(params[:page])
@@ -156,7 +155,7 @@ class Teacher::QuestionnairesController < ApplicationController
     end
 
     def generic_page_params
-      params.require(:generic_page).permit(:course_id, :type_cd, :generic_page_title, :upload_flag,
+      params.require(:generic_page).permit(:course_id, :type_cd, :generic_page_title,
         :file, :start_pass, :start_time, :end_time, :edit_flag,
         :anonymous_flag, :material_memo)
     end
