@@ -103,7 +103,7 @@ Rails.application.configure do
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
   # the I18n.default_locale when a translation cannot be found).
-  config.i18n.fallbacks = true
+  config.i18n.fallbacks = [:en, :ja]
 
   # Send deprecation notices to registered listeners.
   config.active_support.deprecation = :notify
