@@ -45,7 +45,7 @@ RSpec.describe "Users", type: :request do
     end
 
     it 'status OK' do
-      expect(response).to be_success
+      expect(response).to be_successful
       expect(response.status).to eq 200
     end
 
@@ -122,7 +122,7 @@ RSpec.describe "Users", type: :request do
     end
 
     it 'status OK' do
-      expect(response).to be_success
+      expect(response).to be_successful
       expect(response.status).to eq 200
     end
 
